requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0
pydantic[email]==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
langchain==0.1.5
langchain-openai==0.0.5
openai==1.10.0
httpx==0.26.0
python-dotenv==1.0.0
